+++ examples/tutorial/step8.go
@@ -24,7 +24,7 @@ func main() {
 	outputFlows, err := flow.SetSplitter(firstFlow, mySplitter, flowN, nil)
 	flow.CheckFatal(err)
 	flow.CheckFatal(flow.SetStopper(outputFlows[0]))
-	for i := uint8(1); i < flowN; i++ {
+	for i := uint16(1); i < flowN; i++ {
 		flow.CheckFatal(flow.SetHandler(outputFlows[i], modifyPacket[i-1], nil))
 		flow.CheckFatal(flow.SetSender(outputFlows[i], i-1))
 	}
